adjust athena comparison paths
