+++ splink/athena/athena_comparison_level_library.py
@@ -1,85 +1,12 @@
 from ..comparison_level_composition import and_, not_, or_  # noqa: F401
-from ..comparison_level_library import (
-    ArrayIntersectLevelBase,
-    ColumnsReversedLevelBase,
-    DistanceFunctionLevelBase,
-    DistanceInKMLevelBase,
-    ElseLevelBase,
-    ExactMatchLevelBase,
-    LevenshteinLevelBase,
-    NullLevelBase,
-    PercentageDifferenceLevelBase,
+from .athena_helpers.athena_comparison_imports import (  # noqa: F401
+    array_intersect_level,
+    columns_reversed_level,
+    distance_function_level,
+    distance_in_km_level,
+    else_level,
+    exact_match_level,
+    levenshtein_level,
+    null_level,
+    percentage_difference_level,
 )
-from .athena_base import (
-    AthenaBase,
-)
-
-
-class AthenaComparisonProperties(AthenaBase):
-    @property
-    def _exact_match_level(self):
-        return exact_match_level
-
-    @property
-    def _null_level(self):
-        return null_level
-
-    @property
-    def _else_level(self):
-        return else_level
-
-    @property
-    def _array_intersect_level(self):
-        return array_intersect_level
-
-    @property
-    def _columns_reversed_level(self):
-        return columns_reversed_level
-
-    @property
-    def _distance_in_km_level(self):
-        return distance_in_km_level
-
-    @property
-    def _distance_function_level(self):
-        return distance_function_level
-
-    @property
-    def _levenshtein_level(self):
-        return levenshtein_level
-
-
-class null_level(AthenaBase, NullLevelBase):
-    pass
-
-
-class exact_match_level(AthenaBase, ExactMatchLevelBase):
-    pass
-
-
-class else_level(AthenaBase, ElseLevelBase):
-    pass
-
-
-class columns_reversed_level(AthenaBase, ColumnsReversedLevelBase):
-    pass
-
-
-class distance_function_level(AthenaBase, DistanceFunctionLevelBase):
-    pass
-
-
-class levenshtein_level(AthenaBase, LevenshteinLevelBase):
-    pass
-
-
-class array_intersect_level(AthenaBase, ArrayIntersectLevelBase):
-    pass
-
-
-class percentage_difference_level(AthenaBase, PercentageDifferenceLevelBase):
-    pass
-
-
-class distance_in_km_level(AthenaBase, DistanceInKMLevelBase):
-    pass
