+++ splink/athena/athena_linker.py
@@ -8,17 +8,17 @@ import boto3
 import numpy as np
 import pandas as pd
 
-from ..athena.athena_transforms import cast_concat_as_varchar
-from ..athena.athena_utils import (
-    _garbage_collection,
-    _verify_athena_inputs,
-)
 from ..input_column import InputColumn
 from ..linker import Linker
 from ..logging_messages import execute_sql_logging_message_info, log_sql
 from ..misc import ensure_is_list
 from ..splink_dataframe import SplinkDataFrame
 from ..sql_transform import sqlglot_transform_sql
+from .athena_helpers.athena_transforms import cast_concat_as_varchar
+from .athena_helpers.athena_utils import (
+    _garbage_collection,
+    _verify_athena_inputs,
+)
 
 logger = logging.getLogger(__name__)
